fixes for fc40

+++ scripts/fetch_imgtdb.sh
@@ -81,6 +81,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         # Make sed command work also for mac, see: https://stackoverflow.com/a/44864004
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
@@ -95,6 +103,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         # Make sed command work also for mac, see: https://stackoverflow.com/a/44864004
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
@@ -109,6 +125,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         # Make sed command work also for mac, see: https://stackoverflow.com/a/44864004
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
@@ -124,6 +148,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
@@ -137,6 +169,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         # Make sed command work also for mac, see: https://stackoverflow.com/a/44864004
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
@@ -151,6 +191,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
@@ -170,6 +218,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
@@ -184,6 +240,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
@@ -210,6 +274,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
@@ -224,6 +296,14 @@ do
         #echo $URL
         wget2 $URL -O $TMP_FILE -q
         awk '/<pre>/{i++}/<\/pre>/{j++}{if(j==2){exit}}{if(i==2 && j==1 && $0!~"^<pre>"){print}}' $TMP_FILE > $FILE_NAME
+
+        # Check file exists and is not empty
+        if [ ! -s "$FILE_NAME" ]
+        then
+            echo "IMGT Fasta file does not exist, or is empty. Is the IMGT server online?"
+            exit 1
+        fi
+
         sed -i.bak "$REPLACE_VALUE" $FILE_NAME && rm $FILE_NAME.bak
         rm $TMP_FILE
     done
